requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
appnope==0.1.4
arxiv==2.3.1
asttokens==3.0.1
async-timeout==4.0.3
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
comm==0.2.3
dataclasses-json==0.6.7
debugpy==1.8.19
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.1
executing==2.2.1
feedparser==6.0.12
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.7.0
jupyter_core==5.9.1
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.6
langchain-openai==1.1.6
langchain-text-splitters==1.1.0
langsmith==0.6.0
marshmallow==3.26.2
matplotlib-inline==0.2.1
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.2.6
openai==2.14.0
orjson==3.11.5
packaging==25.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
stack-data==0.6.3
tenacity==9.1.2
tiktoken==0.12.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
wcwidth==0.2.14
wikipedia==1.4.0
yarl==1.22.0
zstandard==0.25.0
